requirements.txt
attrs==19.3.0
Babel==0.9.6
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
Beaker==1.5.4
bitarray==2.4.1
bleach==3.1.0
blivet==0.61.15.76
Bottleneck==0.7.0
bpython==0.12
Brlapi==0.6.0
cffi==1.6.0
chardet==2.2.1
configobj==4.7.2
configparser==3.7.4
configshell-fb==1.1.25
contextlib2==0.6.0.post1
coverage==3.6b3
cryptography==1.7.2
cupshelpers==1.0
custodia==0.3.1
Cython==0.19
decorator==3.4.0
defusedxml==0.5.0
di==0.3
dnspython==1.16.0
docutils==0.11
entrypoints==0.3
enum34==1.0.4
ethtool==0.8
fail2ban==0.11.2
fastcache==1.0.2
firstboot==19.5
fros==1.0
functools32==3.2.3.post2
futures==3.1.1
gmpy2==2.0.5
gssapi==1.2.0
gyp==0.1
idna==2.4
importlib-metadata==1.6.0
iniparse==0.4
iotop==0.6
ipaclient==4.6.8
ipaddress==1.0.16
ipalib==4.6.8
ipaplatform==4.6.8
ipapython==4.6.8
IPy==0.75
ipykernel==4.10.0
ipython==3.2.1
ipython-genutils==0.2.0
javapackages==1.0.0
Jinja2==2.7.2
jsonpointer==1.9
jsonschema==3.2.0
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
jwcrypto==0.4.2
kerberos==1.1
kitchen==1.1.1
kmod==0.1
langtable==0.0.31
lxml==3.2.1
M2Crypto==0.21.1
Magic-file-extensions==0.2
Mako==0.8.1
MarkupSafe==0.11
matplotlib==1.2.0
mercurial==2.6.2
mistune==0.8.3
mpmath==1.0.0
mysql-connector-python==1.1.6
MySQL-python==1.2.5
nbconvert==5.4.1
nbformat==4.4.0
netaddr==0.7.5
netifaces==0.10.4
nose==1.3.7
notebook==5.7.8
ntplib==0.3.2
numexpr==2.7.0
numpy==1.13.3
openlmi==0.5.0
pandas==0.17.1
pandocfilters==1.4.2
Paste==1.7.5.1
path.py==5.2
pathlib2==2.3.3
pcp==4.1
perf==0.1
pexpect==2.3
pickleshare==0.7.5
Pillow==2.0.0
ply==3.4
policycoreutils-default-encoding==0.1
prometheus-client==0.6.0
prompt-toolkit==1.0.15
psutil==5.6.7
psycopg2==2.5.1
ptyprocess==0.6.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycups==1.9.63
pycurl==7.19.0
pyglet==1.2.1
Pygments==2.3.1
pygobject==3.22.0
pygpgme==0.3
PyGreSQL==4.0
pyinotify==0.9.4
pykickstart==1.99.66.22
pyliblzma==0.5.3
pyOpenSSL==0.13.1
PyPAM==0.5.0
pyparsing==1.5.6
pyparted==3.9
pyrsistent==0.16.0
pysmbc==1.0.13
python-augeas==0.5.0
python-dateutil==2.8.0
python-dmidecode==3.10.13
python-ldap==2.4.15
python-linux-procfs==0.4.9
python-meh==0.25.3
python-nss==0.16.0
python-yubico==1.2.3
pytz==2016.10
pyudev==0.15
pyusb==1.0.0b1
pywbem==0.7.0
pyxattr==0.5.1
pyxdg==0.25
pyzmq==18.0.1
qrcode==5.0.1
repoze.lru==0.4
requests==2.6.0
rtslib-fb==2.1.74
scandir==1.10.0
scdate==1.10.6
schedutils==0.4
scikit-learn==0.19.0
scipy==0.19.1
Send2Trash==1.5.0
seobject==0.1
sepolicy==1.1
setroubleshoot==1.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
Sphinx==1.1.3
SSSDConfig==1.16.5
stevedore==0.14
subprocess32==3.5.4
sympy==0.7.6
tables==3.2.0rc3
targetcli-fb==2.1.53
Tempita==0.5.1
terminado==0.8.2
Terminator==1.91
testpath==0.4.2
timeout-decorator==0.4.0
tornado==5.1.1
traitlets==4.3.2
urlgrabber==3.10
urllib3==1.10.2
urwid==1.1.1
virtualenv==15.1.0
virtualenv-clone==0.2.4
virtualenvwrapper==4.7.2
wcwidth==0.1.7
webencodings==0.5.1
XStatic==1.0.1
XStatic-jquery-ui==1.12.0.1
yum-langpacks==0.4.2
yum-metadata-parser==1.1.4
zipp==1.2.0
